+++ src/StardewModdingAPI/Events/PlayerEvents.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using StardewModdingAPI.Framework;
-using StardewModdingAPI.Inheritance;
 using StardewValley;
 
 namespace StardewModdingAPI.Events