+++ release-notes.md
@@ -1,5 +1,15 @@
 ﻿# Release notes
 
+## 1.9
+See [log](https://github.com/Pathoschild/SMAPI/compare/1.8...1.9).
+
+For players:
+* Updated for Stardew Valley 1.2.
+
+For mod developers:
+* Many deprecated APIs have been removed; see [deprecation guide](http://canimod.com/guides/updating-a-smapi-mod)
+  for more information.
+
 ## 1.8
 See [log](https://github.com/Pathoschild/SMAPI/compare/1.7...1.8).